pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "psycopg2-binary>=2.9.11",
]
